+++ app/models/variable.rb
@@ -77,7 +77,7 @@ class Variable < ApplicationRecord
   squish :name, :display_name, :field_note, :prepend, :append, :units, :calculated_format
 
   include Translatable
-  translates :display_name, :field_note
+  translates :display_name, :field_note, :prepend, :units, :append
 
   # Scopes
   scope :with_user, ->(arg) { where(user_id: arg) }